functions/func-check/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-file-datalake==12.14.0
python-dotenv==1.0.1
pyodbc
promptflow
promptflow-tools
promptflow-core
promptflow-devkit
promptflow-tracing
promptflow-azure